[improvement] Cast values returned from insert (1h)
Values returned in a `PG::Result` (or an `ActiveRecord::Result`, for that matter) aren't cast using ActiveRecord's type casting. This means that, for example, a custom type-casting for uuids that removed dashes from the string and would be applied for other fetches out of the database (including `.pluck`) wouldn't be applied to an `insert_many` that returned `id`. Therefore, we have to implement that casting ourselves, grabbing the appropriate type and applying it to any returned values so that those values are consistent with their counterparts pulled out of the database in a different way.